// api/wonService/schema.ts
export const CourseSchema = Type.Object({
  publicKey: Type.String(),
  title: Type.String(),
  description: Type.String(),
  syllabus: Type.String(),
  teaser: Type.String(),
  coverArt: Type.String(),
  createdAt: Type.String(),
  publishedAt: Type.String(),
  archivedAt: Type.String(),
})
export type CourseType = Static<typeof CourseSchema>
// ...
export const LessonPlanSchema = Type.Object({
  publicKey: Type.String(),
  courseKey: Type.String(),
  title: Type.String(),
  description: Type.String(),
  objective: Type.String(),
  sequence: Type.Number(),
  coverArt: Type.String(),
  createdAt: Type.String(),
  publishedAt: Type.String(),
  archivedAt: Type.String(),
})
export type LessonPlanType = Static<typeof LessonPlanSchema>

export const LessonContentSchema = Type.Object({
  publicKey: Type.String(),
  lessonKey: Type.String(),
  lessonContentType: Type.String(),
  content: Type.Any(),
  sequence: Type.Number(),
})
export type LessonContentType = Static<typeof LessonContentSchema>
// ...
export const LessonPathSchema = Type.Object({
  publicKey: Type.Optional(Type.String()),
  courseKey: Type.Optional(Type.String()),
  name: Type.Optional(Type.String()),
  description: Type.Optional(Type.String()),
  trailhead: Type.Optional(Type.String()),
})
export type LessonPathType = Static<typeof LessonPathSchema>